+++ src/main/java/com/flansmod/common/network/packets/PacketDriveableSync.java
@@ -42,12 +42,12 @@ public class PacketDriveableSync implements IPacket, IMessage{
 	public void toBytes(ByteBuf bbuf){
 		PacketBuffer buf = new PacketBuffer(bbuf);
 		buf.writeInt(entityId);
-		buf.writeFloat(primary.red);
-		buf.writeFloat(primary.green);
-		buf.writeFloat(primary.blue);
-		buf.writeFloat(secondary.red);
-		buf.writeFloat(secondary.green);
-		buf.writeFloat(secondary.blue);
+		buf.writeByte(primary.red);
+		buf.writeByte(primary.green);
+		buf.writeByte(primary.blue);
+		buf.writeByte(secondary.red);
+		buf.writeByte(secondary.green);
+		buf.writeByte(secondary.blue);
 		buf.writeString(texture_url);
 		buf.writeString(upgrades.toString());
 	}
@@ -56,8 +56,8 @@ public class PacketDriveableSync implements IPacket, IMessage{
 	public void fromBytes(ByteBuf bbuf){
 		PacketBuffer buf = new PacketBuffer(bbuf);
 		entityId = buf.readInt();
-		primary = new RGB(buf.readFloat(), buf.readFloat(), buf.readFloat());
-		secondary = new RGB(buf.readFloat(), buf.readFloat(), buf.readFloat());
+		primary = new RGB(buf.readByte(), buf.readByte(), buf.readByte());
+		secondary = new RGB(buf.readByte(), buf.readByte(), buf.readByte());
 		texture_url = buf.readString(999);
 		upgrades = JsonUtil.getFromString(buf.readString(9999)).getAsJsonArray();
 	}
